[Lite] sync code check

+++ mindspore/lite/src/litert/kernel/cpu/fp32/fill_fp32.cc
@@ -43,7 +43,7 @@ int FillCPUKernel::ReSize() {
   }
   auto output = out_tensors_.front();
   CHECK_NULL_RETURN(output);
-  data_size_ = output->ElementsNum();
+  data_size_ = static_cast<int>(output->ElementsNum());
   thread_sz_count_ = MSMIN(thread_num_, data_size_);
   if (thread_sz_count_ != 0) {
     thread_sz_stride_ = UP_DIV(data_size_, thread_sz_count_);
