+++ mindspore/lite/src/litert/kernel/cpu/fp32/layer_norm_fp32.cc
@@ -40,10 +40,11 @@ int LayerNormCPUKernel::ReSize() {
   auto input = in_tensors_.front();
   CHECK_NULL_RETURN(input);
   auto shape = input->shape();
-  param_->begin_norm_axis_ =
-    param_->begin_norm_axis_ >= 0 ? param_->begin_norm_axis_ : param_->begin_norm_axis_ + shape.size();
-  param_->begin_params_axis_ =
-    param_->begin_params_axis_ >= 0 ? param_->begin_params_axis_ : param_->begin_params_axis_ + shape.size();
+  param_->begin_norm_axis_ = param_->begin_norm_axis_ >= 0 ? param_->begin_norm_axis_
+                                                           : param_->begin_norm_axis_ + static_cast<int>(shape.size());
+  param_->begin_params_axis_ = param_->begin_params_axis_ >= 0
+                                 ? param_->begin_params_axis_
+                                 : param_->begin_params_axis_ + static_cast<int>(shape.size());
   MS_CHECK_LT(param_->begin_norm_axis_, static_cast<int>(shape.size()), RET_ERROR);
   MS_CHECK_LT(param_->begin_params_axis_, static_cast<int>(shape.size()), RET_ERROR);
   param_->norm_outer_size_ = 1;
